Fix for packet upload errors.  This patch is already available in master latest.

+++ registration-processor/init/registration-processor-packet-receiver-stage/src/main/java/io/mosip/registration/processor/packet/receiver/stage/PacketReceiverStage.java
@@ -128,7 +128,14 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 	 * @param routingContext the routing context
 	 */
 	public void failure(RoutingContext routingContext) {
-		this.setResponseWithDigitalSignature(routingContext, globalExceptionHandler.handler(routingContext.failure()), APPLICATION_JSON);
+		try {
+			deleteFile(getFileFromCtx(routingContext));
+		} catch (IOException e) {
+			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
+					"", e.getMessage() + ExceptionUtils.getStackTrace(e));
+		}
+		this.setResponseWithDigitalSignature(routingContext, globalExceptionHandler.handler(routingContext.failure()),
+				APPLICATION_JSON);
 	}
 
 
@@ -153,11 +160,7 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 					"", e.getMessage() + ExceptionUtils.getStackTrace(e));
 			throw new UnexpectedException(e.getMessage());
 		} finally {
-			if (file != null) {
-				if (file.exists()) {
-					deleteFile(file);
-				}
-			}
+			deleteFile(file);
 		 }
 		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
 				"", "PacketReceiverStage::processPacket()::exit");
@@ -203,7 +206,11 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 	 */
 	private void deleteFile(File file) {
 		try {
-			FileUtils.forceDelete(file);
+			if (file != null) {
+				if (file.exists()) {
+					FileUtils.forceDelete(file);
+				}
+			}
 		} catch (IOException e) {
 			throw new UnexpectedException(e.getMessage());
 		}
